vinit changes addded

+++ routes/dataImportExport/excelImportRoutes.js
@@ -6,6 +6,7 @@ const fs = require('fs');
 
 const { importExcel } = require('../../controllers/dataImportExport/ExcelImportController');
 
+const { appendExcel } = require('../../controllers/dataImportExport/AppendExcel');
 // Configure multer to use memory storage
 const upload = multer({ storage: multer.memoryStorage() });
 
@@ -22,6 +23,36 @@ router.post('/import-excel', upload.any(), async (req, res) => {
     fs.writeFileSync(tempPath, file.buffer);
 
     const result = await importExcel(tempPath);
+    
+    // Clean up the temporary file
+    fs.unlinkSync(tempPath);
+
+    if (result.success) {
+      res.json({ message: result.message, details: result.results });
+    } else {
+      res.status(500).json({ error: result.error });
+    }
+  } catch (error) {
+    console.error('Error in import-excel route:', error);
+    res.status(500).json({ error: 'An unexpected error occurred' });
+  }
+});
+
+
+
+router.post('/append-excel', upload.any(), async (req, res) => {
+  try {
+    if (!req.files || req.files.length === 0) {
+      return res.status(400).send('No file uploaded.');
+    }
+
+    const file = req.files[0]; // Get the first file
+    const tempPath = path.join(__dirname, '../../uploads', file.originalname);
+
+    // Write the file to disk
+    fs.writeFileSync(tempPath, file.buffer);
+
+    const result = await appendExcel(tempPath);
     
     // Clean up the temporary file
     fs.unlinkSync(tempPath);
